[testutils] move the non-volatile flash counter funcs to separate lib
The non-volatile flash counters testutils were placed in the
`flash_ctrl_testutils` lib. However, this required all programs that use
these counter to have special symbols that are only put in place by the
flash linker script. #17389 needs to be able to use some of the
`flash_ctrl_testutils` functions that do not touch the non-volatile
counters in an SRAM program that does not define such symbols. To fix
this issue, a separate testutils lib is created to hold the non-volatile
counter functions.
